voi pyhä lehmä...eli appBarille tein uuden componentin ja muutos vanhaan SearchBariin on...en tallentanut datasourcee stateen...ja sit se toimi
