+++ src/App.js
@@ -7,7 +7,6 @@ import Activities from './components/Activities';
 import Appbar from './components/AppBar';
 import activitiesData from './partio.json';
 import { API_ROOT } from './api-config';
-import { filterOffExistingOnes} from './functions/searchBarFilter'
 
 class App extends Component {
   constructor() {
@@ -75,32 +74,20 @@ class App extends Component {
           <div id="container">
             <div className="content">
               <h2 style={{ marginTop: 120 }}>Events</h2>
-              <NewEvent updateEvents={this.updateEvents} />
+             <NewEvent updateEvents={this.updateEvents} />
               <ListEvents
                 events={this.state.events}
                 fetchEvents={this.getEvents}
                 fetchedActivities={this.state.activities}
               />
-            
-            
-            
-            
-            
-             <p> {filterOffExistingOnes(this.state.activities, this.state.events, null)}
-           </p>
-           
-           
-           
-           
-           
-           
               <h2>Activities</h2>
               <Activities fetchedActivities={this.state.activities} />
             </div>
             <Sticky>
               {({ style }) => (
                 <header style={style}>
-                  <Appbar activities={this.state.activities} />
+                  <Appbar activities={this.state.activities}
+                  events={this.state.events} />
                 </header>
               )}
             </Sticky>
