+++ src/components/AppBar.js
@@ -1,17 +1,20 @@
 import React from 'react';
-import ActivitySearch from './SearchBar';
-import { activitiesArray } from './Activities';
-
+import TopSearchBar from './TopSearchBar';
+import {filterOffExistingOnes} from '../functions/searchBarFiltering';
 /**
  * A simple example of `AppBar` with an icon on the right.
  * By default, the left icon is a navigation-menu.
  */
-const Appbar = ({ activities }) => (
+const Appbar = ({ activities, events }) => {
   // Parametreina tulee vielä bufferzone drag and droppia varten
 
+  const filteredActivities = filterOffExistingOnes(activities, events)
+  console.log(filteredActivities)
+  return (
   <div className="top-search" style={{background: '#5DBCD2', padding: 20}}>
-    <ActivitySearch dataSource={activitiesArray(activities)} />
+    <TopSearchBar dataSource={filteredActivities} />
   </div>
-);
+  )
+};
 
 export default Appbar;
