+++ src/components/TopSearchBar.js
@@ -5,49 +5,14 @@ import Select from 'react-select';
 import 'react-select/dist/react-select.css';
 import { API_ROOT } from '../api-config';
 
-export default class ActivitySearch extends React.Component {
+export default class TopSearchBar extends React.Component {
   constructor(props) {
     super(props);
+    console.log(this.props.dataSource)
     this.state = {
-     dataSource: this.props.dataSource,
       selectedActivity: null
     };
   }
-
-  updateSelectedActivity = selectedActivity => {
-    console.log('Update selected activity', selectedActivity);
-    this.setState({
-      selectedActivity
-    });
-  };
-
-  saveActivityToEvent = () => {
-    if (this.state.selectedActivity) {
-      const data = {
-        guid: this.state.selectedActivity.value
-      };
-
-      console.log('Tallenna aktiviteetti', data);
-
-      fetch(`${API_ROOT}/events/${this.props.event.id}/activities`,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(data)
-        }
-      )
-        .then(res => res.json())
-        .then(res => this.props.updateActivities(res))
-        .then(
-          this.setState({
-            selectedActivity: null
-          })
-        )
-        .catch(error => console.error('Error:', error));
-    } else {
-      console.log('Ei valittua aktiviteettia');
-    }
-  };
 
   handleChange = selectedActivity => {
     this.setState({ selectedActivity });
@@ -57,6 +22,7 @@ export default class ActivitySearch extends React.Component {
   render() {
     const { selectedActivity } = this.state;
     const value = selectedActivity && selectedActivity.value;
+    console.log(this.props.dataSource)
     return (
       <div>
         <Select
@@ -67,7 +33,7 @@ export default class ActivitySearch extends React.Component {
             const sorterOptions = { keys: ['label'] };
             return matchSorter(options, filter, sorterOptions);
           }}         
-          options={this.state.dataSource.map(activity => {
+          options={this.props.dataSource.map(activity => {
             let obj = {};
             obj = { value: activity.guid, label: activity.title };
             return obj;
@@ -84,7 +50,7 @@ export default class ActivitySearch extends React.Component {
           <RaisedButton
             label="Tallenna aktiviteetti"
             primary={true}
-            onClick={this.saveActivityToEvent}
+            onClick={console.log('click')}
           />
         </div>
       </div>
